api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.7
pgvector>=0.3
python-multipart>=0.0.9
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
pdfplumber>=0.11
trafilatura>=1.12
aiofiles>=24.1
slowapi>=0.1.9
sentence-transformers>=3.0
torch>=2.0
